added Logger

+++ src/main/java/com/college/student/crudservlet/DeleteStudentServlet.java
@@ -12,21 +12,37 @@ import java.io.PrintWriter;
 public class DeleteStudentServlet extends HttpServlet {
     public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
         response.setContentType("text/html");
-        StudentService studentService = (StudentService) request.getServletContext().getAttribute("studentService");
-        Student student = studentService.deleteStudentByRollNo(Integer.parseInt(request.getParameter("rollNo")));
         PrintWriter out = response.getWriter();
-        out.println("<!DOCTYPE html>");
-        out.println("<html lang=\"en\">");
-        out.println("<head>");
-        out.println("<meta charset=\"UTF-8\">");
-        out.println("<meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">");
-        out.println("<title>Added</title>");
-        out.println("</head>");
-        out.println("<body>");
-        out.println("<h1>" + student.getRollNo() + "   " + student.getName() + "  " + student.getRollNo() + "  " + student.getPhoneNo() + "<h1>");
-        out.println("<h1>Student Detail's Deleted Successfully</h1>");
-        out.println("<a href='StudentChoices.html'><h1>Main Page<h1></a>");
-        out.println("</body>");
-        out.println("</html>");
+        StudentService studentService = (StudentService) request.getServletContext().getAttribute("studentService");
+        if (!studentService.isStudentExist(Integer.parseInt(request.getParameter("rollNo")))) {
+            out.println("<!DOCTYPE html>");
+            out.println("<html lang=\"en\">");
+            out.println("<head>");
+            out.println("<meta charset=\"UTF-8\">");
+            out.println("<meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">");
+            out.println("<title>notFound</title>");
+            out.println("</head>");
+            out.println("<body>");
+            out.println("<h3>Student with rollNo :  " + request.getParameter("rollNo") + " <strong> - Not Found</strong><h3>");
+            out.println("click below to manage students<br>");
+            out.println("<br>click below : <a href='StudentChoices.html'><h2>click here<h2></a>");
+            out.println("</body>");
+            out.println("</html>");
+        } else {
+            Student student = studentService.deleteStudentByRollNo(Integer.parseInt(request.getParameter("rollNo")));
+            out.println("<!DOCTYPE html>");
+            out.println("<html lang=\"en\">");
+            out.println("<head>");
+            out.println("<meta charset=\"UTF-8\">");
+            out.println("<meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">");
+            out.println("<title>Added</title>");
+            out.println("</head>");
+            out.println("<body>");
+            out.println("<h1>" + student.getRollNo() + "   " + student.getName() + "  " + student.getRollNo() + "  " + student.getPhoneNo() + "<h1>");
+            out.println("<h1>Student Detail's Deleted Successfully</h1>");
+            out.println(" click here : <a href='StudentChoices.html'><h1>Main Page<h1></a>");
+            out.println("</body>");
+            out.println("</html>");
+        }
     }
 }
